requirements.txt
# Web framework
flask==3.1.0
flask-cors==4.0.1

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10

# Auth
authlib==1.3.2
PyJWT==2.9.0
cryptography==43.0.3

# Validation
pydantic==2.9.2

# Scheduling
APScheduler==3.10.4

# Utils
python-dotenv==1.0.1
requests==2.32.3

# Production server
gunicorn==23.0.0
